Pi/requirements.txt
flask
flask-cors
Adafruit_WS2801
RPi.GPIO